voice_note/server/stt/requirements.txt
PyAudio==0.2.14
torchaudio==2.4.0
transformers==4.44.0
